requirements.txt
pynput>=1.8.1
mutagen>=1.47.0
PySide6>=6.9.0
psutil>=7.0.0
requests>=2.32.3
packaging>=25.0
python-dateutil>=2.9.0
ffmpeg-python>=0.2.0
yt-dlp>=2025.8.27